refactor(communities): remove legacy pending membership indicator
We no longer want to show pending membership requests in the left column
of the layout. Instead, community owners need to go to the community
settings -> members -> Pending requests

This however does not remove the legacy membership popup yet because
it's still used when users click the OS notification of incoming
community requests to join.

We don't yet have a mechanism to make the app navigate specifically to
that section of the community settings.