fix(Wallet) move deletion of pending transactions in nim
Bump status-go HEAD that removes deletion of pending multi-transactions
in status-go

Previous implementation was relying on the fact that all pending
transactions are also multi-transactions, which it seems not to be
the case (e.g. ens, airdrop collectible)

Also tried to extend the deletion of multi-transactions to status go
but it was causing crashes for minted/dropped tokens.

Updates #10474